# docs/openapi.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"AuthController_register","summary":"Đăng ký tài khoản demo","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"Đăng nhập","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/procedures":{"get":{"operationId":"ProceduresController_findAll","summary":"Danh sách thủ tục hành chính","parameters":[],"responses":{"200":{"description":""}},"tags":["Procedures"]}},"/procedures/{code}":{"get":{"operationId":"ProceduresController_findOne","summary":"Chi tiết thủ tục + checklist giấy tờ","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Procedures"]}},"/procedures/identify":{"post":{"operationId":"ProceduresController_identify","summary":"HoSoBot — nhận diện thủ tục từ câu hỏi tự nhiên (Bước 1)","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdentifyDto"}}}},"responses":{"201":{"description":""}},"tags":["Procedures"]}},"/document-types":{"get":{"operationId":"DocumentTypesController_findAll","summary":"Danh mục loại giấy tờ (catalog dùng chung)","parameters":[],"responses":{"200":{"description":""}},"tags":["DocumentTypes"]}},"/document-types/{code}":{"get":{"operationId":"DocumentTypesController_findOne","summary":"Chi tiết loại giấy tờ + field schema","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["DocumentTypes"]}},"/documents/upload":{"post":{"operationId":"DocumentsController_upload","summary":"Upload giấy tờ (Bước 2)","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadFileDto"}}}},"responses":{"201":{"description":""}},"tags":["Documents"]}},"/documents/{sessionId}/ocr/{documentTypeCode}":{"post":{"operationId":"DocumentsController_triggerOcr","summary":"Trigger OCR cho giấy tờ (Bước 3) — gọi VNPT API qua ai-svc","parameters":[{"name":"sessionId","required":true,"in":"path","schema":{"type":"string"}},{"name":"documentTypeCode","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TriggerOcrDto"}}}},"responses":{"201":{"description":""}},"tags":["Documents"]}},"/sessions/{id}/crosscheck":{"post":{"operationId":"ScoringController_crosscheck","summary":"CrossCheck đối chiếu chéo giấy tờ (Bước 4)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Scoring"]}},"/sessions/{id}/score":{"post":{"operationId":"ScoringController_score","summary":"Chấm điểm hồ sơ 0–100 (Bước 5)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Scoring"]}},"/sessions/{id}/lawguard":{"post":{"operationId":"ScoringController_lawguard","summary":"LawGuard — tra cứu căn cứ pháp lý RAG (Bước 6)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Scoring"]}},"/sessions":{"post":{"operationId":"SessionsController_create","summary":"Tạo phiên kiểm tra hồ sơ mới","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSessionDto"}}}},"responses":{"201":{"description":""}},"tags":["Sessions"]}},"/sessions/{id}":{"get":{"operationId":"SessionsController_findOne","summary":"Lấy trạng thái phiên (dùng cho progress bar)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Sessions"]}},"/sessions/{id}/confirm":{"post":{"operationId":"SessionsController_confirm","summary":"Người dân xác nhận hồ sơ (Bước 8)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Sessions"]}},"/sessions/{id}/smartform":{"post":{"operationId":"SmartFormController_generate","summary":"SmartForm — tự điền form từ dữ liệu OCR (Bước 7)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["SmartForm"]}},"/sessions/{id}/recheck":{"post":{"operationId":"RecheckController_recheck","summary":"Gov Re-Check — cán bộ tái kiểm hồ sơ, phát hiện riskFlags (Bước 9)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Recheck"],"security":[{"bearer":[]}]}},"/priority":{"get":{"operationId":"PriorityController_getQueue","summary":"Danh sách hồ sơ xếp theo ưu tiên A/B/C/D + SLA (Bước 10)","parameters":[],"responses":{"200":{"description":""}},"tags":["Priority"],"security":[{"bearer":[]}]}},"/insights/dashboard":{"get":{"operationId":"InsightsController_dashboard","summary":"InsightMap Dashboard — tổng hợp điểm nghẽn (Bước 11)","parameters":[{"name":"days","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Insights"],"security":[{"bearer":[]}]}},"/insights/top-errors":{"get":{"operationId":"InsightsController_topErrors","summary":"Top lỗi theo loại (metadata ẩn danh)","parameters":[{"name":"days","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Insights"],"security":[{"bearer":[]}]}},"/insights/trend":{"get":{"operationId":"InsightsController_trend","summary":"Xu hướng lỗi theo ngày","parameters":[{"name":"days","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Insights"],"security":[{"bearer":[]}]}}},"info":{"title":"GovTrust AI API","description":"API Orchestrator — Hệ thống tiền kiểm hồ sơ dịch vụ công","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"username":{"type":"string","example":"nguyenvana","description":"Tên đăng nhập"},"password":{"type":"string","example":"matkhau123","minLength":6,"description":"Mật khẩu (>= 6 ký tự)"},"fullName":{"type":"string","example":"Nguyễn Văn A","description":"Họ tên đầy đủ"},"role":{"type":"string","enum":["CITIZEN","OFFICER","ADMIN"],"example":"CITIZEN","description":"Vai trò; mặc định CITIZEN. Đăng ký OFFICER để test Bước 9–11."}},"required":["username","password","fullName"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string","example":"nguyenvana"},"password":{"type":"string","example":"matkhau123"}},"required":["username","password"]},"IdentifyDto":{"type":"object","properties":{"userQuery":{"type":"string","example":"Tôi muốn đăng ký khai sinh cho con","description":"Câu hỏi tự nhiên của người dân để HoSoBot nhận diện thủ tục"}},"required":["userQuery"]},"UploadFileDto":{"type":"object","properties":{"sessionId":{"type":"string","example":"665f1b2c9a1e4a0012345678","description":"ID phiên kiểm tra"},"documentTypeCode":{"type":"string","example":"CCCD","description":"Mã loại giấy tờ (CCCD, GIAY_KHAI_SINH, ...)"},"file":{"type":"string","format":"binary","description":"File ảnh/PDF giấy tờ"}},"required":["sessionId","documentTypeCode","file"]},"TriggerOcrDto":{"type":"object","properties":{"checklistId":{"type":"string","example":"cccd_cha_me","description":"Slot checklist mà giấy này điền vào; mặc định = documentTypeCode"}}},"CreateSessionDto":{"type":"object","properties":{"procedureId":{"type":"string","example":"665f1b2c9a1e4a0012345678","description":"ObjectId của thủ tục (procedure) người dân đã chọn"}},"required":["procedureId"]}}}}